chore: Move default excludes to stand-alone file
This allows documentation to provide a permanent link to the master
version of the default exclude list.
